refactor: make integration_test.yaml not depend on PR event (#736)
* feat: make integration_test.yaml not depend on PR event

This commit changes the logic for checking whether there were code changes;
instead of using the PR event exclusively for getting commit SHAs to compare
changes between the PR and the target branch, we use tj-actions/changed-files.
This change will allow this workflow to be reusable outside of the context of a PR,
as well as avoid issues with depending on the commit hashes of the target and current branches.

Fixes #733

* refactor: use custom logic instead of action

* docs: correct changelog